use glob import & assort structure
