+++ src/common/res.ts
@@ -1,0 +1,3 @@
+export const Documents = import.meta.glob('%/**/*.*', { as: 'raw' });
+// export const Definitions = import.meta.glob('~/**/*.*', { as: 'raw '});
+
