+++ src/components/markdown/base/common.ts
@@ -1,4 +1,8 @@
-import { generalIsCodeBlockCategory, generalIsContainer, generalNeedsInlineParse, GeneralNodeType, NodeTypeDefinition } from 'commonmark';
+import { BlockParsingOptions, compileMaybeSpecialRegExp, generalIsCodeBlockCategory, generalIsContainer, generalNeedsInlineParse, GeneralNodeType, NodeTypeDefinition } from 'commonmark';
+import { parseInlineEmoji } from './emoji';
+import { MathHandler, MathTrigger, parseInlineMathFence } from './math';
+import { TableHandler, TableHeadHandler, TableRowHandler, TableCellHandler, TableTrigger } from './table';
+import { parseInlineTemplate } from './template';
 
 export type ExtendedNodeType = GeneralNodeType 
   | 'table' | 'table_row' | 'table_head' | 'table_cell' 
@@ -26,4 +30,28 @@ export const ExtendedNodeDefinition: NodeTypeDefinition<ExtendedNodeType> = {
   needsInlineParse: function (t: ExtendedNodeType): boolean {
     return generalNeedsInlineParse(t) || t === 'table_cell';
   }
+};
+
+
+export const ExtendedSyntaxOptions: BlockParsingOptions<ExtendedNodeType> = {
+  type: ExtendedNodeDefinition,
+  blockHandlers: {
+    table: TableHandler,
+    table_head: TableHeadHandler,
+    table_row: TableRowHandler,
+    table_cell: TableCellHandler,
+  
+    math_block: MathHandler,
+  },
+  blockStartHandlers: {
+    [-1]: [TableTrigger],
+    [2]: [MathTrigger],
+  },
+  reMaybeSpecial: compileMaybeSpecialRegExp('$', '|', true),
+  // reNonSpecialChars: compileNonSpecialCharRegExp('$', true),
+  inlineHandlers: [
+    ['$', parseInlineMathFence], 
+    ['@', parseInlineTemplate],
+    [':', parseInlineEmoji],
+  ]
 };
