+++ src/components/markdown/node/MathBlock.tsx
@@ -1,6 +1,6 @@
 import { MathComponent } from 'mathjax-react';
 import React, { PropsWithChildren } from 'react';
-import { filterStringChildren } from './common';
+import { filterStringChildren } from '../html-renderer';
 
 export const MathSpan = (props: PropsWithChildren<object>) => {
   const tex = filterStringChildren(props.children);
