+++ tsconfig.json
@@ -18,7 +18,8 @@
     "baseUrl": "./src",
     "paths": {
       "@/*": ["./*"],
-      "~/*": ["../defs/*"]
+      "~/*": ["../defs/*"],
+      "%/*": ["../docs/*"]
     },
   },
   "include": ["src", "docs"],
