+++ vite.config.ts
@@ -7,15 +7,16 @@ import { viteStaticCopy as copy } from 'vite-plugin-static-copy';
 export default defineConfig({
   plugins: [
     react(),
-    copy({ targets: [
-      { src: 'docs/**/*', dest: 'md' }
-    ], flatten: false })
+    // copy({ targets: [
+    //   { src: 'docs/**/*', dest: 'md' }
+    // ], flatten: false })
   ],
   resolve: {
     alias: {
       '@': path.resolve(__dirname, './src'),
       '~': path.resolve(__dirname, './defs'),
+      '%': path.resolve(__dirname, './docs'),
     },
   },
-  assetsInclude: ['./docs/**/*.md']
+  // assetsInclude: ['./docs/**/*.md']
 });
